Add gather transactions route and controller
Split the accounts controller to gather transactions and accounts. The gather transactions calls upon the worker to gather bank transactions
and responds with either still working, error or complete.

Update citizen flow and route to incorporate the above
